+++ examples/dts-inspector/CHANGELOG.md
@@ -1,5 +1,19 @@
 # dts-inspector
 
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [faf1852]
+  - @openfn/describe-package@0.0.18
+
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [4c875b3]
+  - @openfn/describe-package@0.0.17
+
 ## 1.0.13
 
 ### Patch Changes